refacto animation + rules
